requirements.txt
adamp
faiss-cpu
ipython==8.12.3
matplotlib==3.6.3
mmcv-full==1.7.1
mmengine==0.9.1
opencv-python
Pillow==9.3.0
pycocotools
ftfy
regex
scikit-learn==1.2.2
scipy==1.10.0
seaborn==0.12.2
timm==0.5.4
tqdm
wandb
yacs
yapf
zipp